\name{sim_scRNA_data}
\docType{data}
\alias{sim_scRNA_data}
\title{A simulation scRNA-seq dataset from splatter package.}
\description{
  This dataset is a SingleCellExperiment class, it includes 5,000 genes and 
  100 cells.
}
\usage{sim_scRNA_data}

\format{A SingleCellExperiment class data includes 5,000 genes and 
  100 cells.}
\source{
  A simulation data generated from spaltter package.
}
\keyword{datasets}

